requirements.txt
Flask
gunicorn
pylint
google-cloud-storage